Bug: Don’t load the environment variables in the cloud!
This code that was intended for use in local development and testing was running in AWS Lambda, and the code ended up thinking that it was in the ‘dev’ environment.  It was very confusing to understand why it kept trying to connect to the development S3 bucket.  TODO: The bug couldn’t have happened if the samconfig.toml file hadn’t been built into the package for Lambda.  It would be better if only the code needed for runtime were included in the package.