OutputRegion, more memory.
The source bucket is specified in the event, and the code doesn’t need to know anything else about it.  So, there is no need to set a source bucket variable.  Also: There IS a need to set an output region, since the output bucket might not be in us-west-1.  (It was not, for a bucket involved in our production use case.)  Also, ImageMagick needed more memory.  TODO: Hopefully we can add a libvips layer soon and then we’ll need less memory for processing large images.